+++ src/components/micro/floating-header/styled-header.js
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import { getThemeProperty } from "@assets/utils/styled";
+import { getThemeProperty } from "@helpers/utilities";
 
 export const HeaderStyled = styled.header`
 position: sticky;
