+++ src/components/micro/input-short-text/styled-input.js
@@ -1,5 +1,5 @@
 import styled, { css } from "styled-components";
-import { getThemeProperty } from "../../../assets/utils/styled";
+import { getThemeProperty } from "@helpers/utilities";
 
 const BaseInputStyle = css`
 padding: .75rem 1rem;
