+++ src/styles/global-styles.js
@@ -1,5 +1,5 @@
 import { createGlobalStyle } from "styled-components";
-import { getThemeProperty } from "@assets/utils/styled";
+import { getThemeProperty } from "@helpers/utilities";
 
 export const GlobalStyle = createGlobalStyle`
 * {
